// src/app/resolvingoferrors/models/solicited-errors.ts
export interface SolicitedErrors {
    View: string;
    TelNo: string;
    Cmd: string;
    Source: string;
    Created: string;
    Status: string;
    ResType: string;
    ErrorList: string;
    Reference: string;
    LatestUserCmt: string;
    LatestCmtDate: string;
    TranId: string; 
    // Ovd: string;  
}